split packages, drop godotenv
